feat(version, cmd/gpud/update): separate DetectLatestVersion function, move "update" to a separate dir
